Bug 909839 - Merge JavaScript 'atoms' used by gaia and gaiatest. r=jgriffin
