+++ tests/python/gaia-ui-tests/gaiatest/.gitignore
@@ -1,0 +1,1 @@
+atoms
