--- tests/python/gaia-ui-tests/gaiatest/atoms/remote_date.js
@@ -1,15 +1,0 @@
-return (function(ms) {
-
-  var date = new Date();
-
-  return {
-    year: date.getFullYear(),
-    month: date.getMonth(),
-    date: date.getDate(),
-    hours: date.getHours(),
-    minutes: date.getMinutes(),
-    seconds: date.getSeconds()
-  };
-
-}.apply(this, arguments));
-
